+++ golem-worker-executor-base/src/durable_host/serialized.rs
@@ -592,10 +592,11 @@ mod tests {
             template_id,
             worker_name,
         })
+
     }
 
     fn promiseid_strat() -> impl Strategy<Value = PromiseId> {
-        (workerid_strat(), any::<i32>()).prop_map(|(worker_id, oplog_idx)| PromiseId {
+        (workerid_strat(), any::<u64>()).prop_map(|(worker_id, oplog_idx)| PromiseId {
             worker_id,
             oplog_idx,
         })
