+++ golem-worker-executor/src/context.rs
@@ -97,7 +97,7 @@ impl ExternalOperations<Context> for Context {
     async fn get_worker_retry_count<T: HasAll<Context> + Send + Sync>(
         this: &T,
         worker_id: &WorkerId,
-    ) -> u32 {
+    ) -> u64 {
         DurableWorkerCtx::<Context>::get_worker_retry_count(this, worker_id).await
     }
